backend/requirements.txt
fastapi
uvicorn
sumy
nltk
pydantic
google-generativeai
python-dotenv
langdetect
scikit-learn
datasets
numpy
imbalanced-learn
